Refactor transiition label provider
